Revert "Revert "review 수정""
This reverts commit b81ea5c267f7a45a77630b78114166935a2d468a.

+++ ToDo/app/src/main/java/com/jmp/todo/model/ImageFileManager.java
@@ -4,13 +4,14 @@ import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
-import android.widget.Toast;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.UUID;
 
 public class ImageFileManager {
     Context context;
@@ -18,14 +19,14 @@ public class ImageFileManager {
         this.context = context;
     }
 
-    public void writeToInternalStorage(Task task) {
+    public String writeToInternalStorage(String imageContent) {
+        String imageName = createName();
         try {
-            if (fileExist(task)) {
-                return;
+            if (isFileExist(imageContent)) {
+                return imageContent;
             }
-            FileOutputStream fos = null;
-            fos = context.openFileOutput(getName(task.getImageContent()), Context.MODE_PRIVATE);
-            InputStream is = context.getContentResolver().openInputStream(Uri.parse(task.getImageContent()));
+            FileOutputStream fos = context.openFileOutput(imageName, Context.MODE_PRIVATE);
+            InputStream is = context.getContentResolver().openInputStream(Uri.parse(imageContent));
             Bitmap image = BitmapFactory.decodeStream(is);
             image.compress(Bitmap.CompressFormat.JPEG, 100, fos);
 
@@ -36,35 +37,36 @@ public class ImageFileManager {
         } catch (IOException e) {
             e.printStackTrace();
         }
+        return imageName;
     }
-    public String getPath(Task task) {
+    public String getPath(String imageContent) {
         try{
             File directory = context.getFilesDir();
-            File image = new File(directory, getName(task.getImageContent()));
+            File image = new File(directory, imageContent);
             return image.toString();
         } catch (NullPointerException e){
             e.printStackTrace();
         }
         return "null";
     }
-    public void deleteImage(Task task) {
-        try{
-            File directory = context.getFilesDir();
-            File image = new File(directory, getName(task.getImageContent()));
-            image.delete();
-        } catch (NullPointerException e){
-            e.printStackTrace();
+    public void deleteDoneImage(ArrayList<Task> tasks) {
+        for (Task task : tasks) {
+            if (!task.isDone()) continue;
+            try{
+                File directory = context.getFilesDir();
+                File image = new File(directory, task.getImageContent());
+                image.delete();
+            } catch (NullPointerException e){
+                e.printStackTrace();
+            }
         }
     }
-    public String getName (String imageContent) {
-        String name = imageContent.replace(":","");
-        name = name.replace("/","");
-        name = name.replace(".","") + ".jpg";
-        return name;
+    public String createName() {
+        return UUID.randomUUID().toString().replace("-","") + ".jpg";
     }
-    public Boolean fileExist(Task task) {
+    public Boolean isFileExist(String imageContent) {
         File directory = context.getFilesDir();
-        File image = new File(directory, getName(task.getImageContent()));
+        File image = new File(directory, imageContent);
         if (image.exists()) {
             return true;
         } else {
